[sw-sysemu] [BEMU] Link emu (libemu.so) as a dynamic library
The emu code is compiled into libchecker.so and sys_emu directly, and
libemu.so/libchecker.so is linked at compile time with the rtl monitors. This
removes the need to dynamically lookup function pointers, and it fixes a few
bugs where the code was attemping to use functions that did not exist, namely
c_jalr() and gfprintf().


Former-commit-id: 600e85659b690d3b75e04082c81a4cc842eb0e60 [formerly 5cb91bd7d8fad81d8a2c505c60221b2ed3f2bbd8]
Former-commit-id: c455df3f1cb3ceb6a6e80607ab1e3052c5d89cd7