[sw-sysemu] [SW-166] Create libsysemu
Create libsysemu that is going to be consumed by the SW repos to
build the final system simulator. Simple changes introduced to make
sure that the main function is not included in libsysemu.a

Create sysemu-artifacts.tar.gz file that includes the sysemu binary,
libsysemu.a and all sysemu headers we could potentially. The list
of headers would probably have to be revisited.


Former-commit-id: ca0c6bcf605d0ebbf4e99469e20ee2c4b278727f