[sw-sysemu] [SW-170] Refactor SysEmu to facilitate integration with the DeviceAPI
Create the sys_emu_cmd_options struct and use that store all
variables used for strong the result of parsing the command
line arguments.

This is in preparation for splitting the main function in the
part that parses the command line arguments and the main exeuction
loop.

Also add the sys_emu.h header that will hold the interface
of libsysemu.

Add function parse_command_line_arguments that returns
a sys_emu_cmd_options struct with the expected values

Introduce the sys_emu class. This is in an effort
to encapsulate in a class all static state defined in the sys_emu.cpp
file.

This class will allow us in the future to create subclasses of the
simulator where we will be able to specialize and overwrite specific
parts of interest: e.g. replace the api we use to talk to the
simulators.

A number of member fields and methods of the classes had to be made
static for now as they are being used in other parts of the code where
they externally referenced (main_memory_region_esr.cpp and
sys_emu_log.cpp) This broken design should be fixed in the future.

Introduces in the sys_emu class the virtual function
allocate_api_listener that will be overwritten by versions of the
simulator that implement a different API than the current one.

We have also introduced the init_simulator function where we move
there the simulator initialization.

Given that we want to be able to overwrite the allocate_api_listener
function during initialization of a subclass, this cannot be done
inside the constructor (an overloaded function cannot be called inside
the base constructor)

Finally make changes to copy a number of headers. This is a hack we
should have a "proper" package installation,
for now just copy the headers and try to maintain the original
directory structure so it is easy to setup the include path