[esperanto-tools-libs] [BUILD] Add the HOST_SW CMake configuration option and fix elf_to_mem.
* Disable building the sdk as part of the runtime on host-sw
* Enable building the host-sw repo
* Update esperanto-soc commit
* Fix path to elf_to_mem artifact script
* Do not compile sdk in runtime compilation

Add the HOST_SW cmake configuration option. This option on by default as
we are building as part of the host-sw repo

This option is added both at the project and on the runtime CMakeFiles.
This is to handle the case were the runtime is built as an external
project by the backend.