[esperanto-tools-libs] [RT][SW-205] Move etrt under the c-api folder
The etrt.h file presents a C-API of the runtime library
Move to the c-api folder where it is going to be presented to any
C like code.

This task is in effort of evolving the Runtime-API to a c++ centric
one that following C++-good-practices and a C oriented one.