[esperanto-tools-libs] [RT][SW-205] Introduce CMakeLists.txt file inside src folder.
Break the top-level CMakeLists.txt file of runtime and move any logic
specific to building the runtime library inside the src folder.
This step is in preparation of further splitting the src folder
into different subfoldes where we split the files based on the
respective libraries they are going to belong.