[sw-sysemu] [tbox] Fxed VK_FORMAT_BC1_RGB_UNORM_BLOCK and VK_FORMAT_BC1_RGBA_UNORM_BLOCK
The VK_FORMAT_BC1_RGB_UNORM_BLOCK and VK_FORMAT_BC1_RGBA_UNORM_BLOCK
formats were not being checked against TBoxEmu. I have updated the scoreboard
for comparing against TBoxEmu and fixed the mismatches I have seen, whether
the underlying issues laid on the TBoxEmu or the RTL side.

For judging whether the RTL or TBoxEmu were producing the right result for
the mismatching outputs I used DirectXTex, which is a publicly available
library for encoding and decoding all the formats supported in DirectX
12, including the BC1 formats that were included in Vulkan.

On the TBoxEmu side the VK_FORMAT_BC1_RGB_UNORM_BLOCK format did produce
transparent texels, even though the spec states that it is an opaque
format. For handling the case when BC1 should produce opaque black texels
I have added the hasAlpha extra parameter to the decode_BC1 function.

On the RTL side the treatment of the opaque texels for
VK_FORMAT_BC1_RGB_UNORM_BLOCK was fine, but we were doing an incorrect
check for how to interpolate. The spec states that there are two sets of
interpolation formulas, picking one or another depending on whether
"c0 > c1", with c0 an c1 being 16-bits values packing 5 bits for red on
the LSB, 6 bits for green and 5 bits for blue. For some reason we did
implement a non-standard check like:

red0 > red1 && green0 > green1 && blue0 > blue1

This check was just wrong and does not relate to the Khronos spec.

A second issuee with the RTL implementation was the way that the 5
and 6 bits components were interpolated. The components were right-padded
with zeros for feeding into the 8 bits interpolation circuitry, as in
this expression:

color0 = {alpha[7:0], color[4:0],3'b0, color[10:5],2'b0, color[15:11],3'b0};

While this expression produces 8 bits, it does not implement a linear
mapping from 5 or 6 bits to the [0,255] interval.

Conceptually, for the red component (color[4:0] in the expression
above), we should do:

unorm8 =  uint8_t( float(red) / 31.f * 255.f + 0.0000001f );

Using floating point one should be able to map 31 to 255, rather than
to 248, which falls 7 units shorter from covering the whole range of
available 8 bits values.

Said in other words, converting to 8 bits by right-padding with zeros
is equivalent to multipliying by 8, when the actual value we should
use is 255.f / 31.f ~ 8.226

For the 5 bits cases I have compensated this inaccuracy by scaling the
iterpolation weights with the "1 / 8.f * 255.f / 31.f" factor. Similarly
for the 6 bits case I scale the interpolation weights with
"1 / 4.f * 255.f / 63.f".

Because of this extra scaling some extra bits of accuracty were needed
for the interpolation weight (they are up to 16 bits, from the previous
11 bits). The needed accuracy was estimated by trial and error as the
mininimum accuracy needed for fully matching the RTL and TBoxEmu.

One last problem this patch fixes is the handling of ties in the
rounding when converting to unorm8. The RTL needed to get an epsilon
value added for matching the TBoxEmu behaviour.

Similarly, the TBoxEmu also needed a small epsilon for handling cases
like 0x3effffff (0.49999997) when converting to unorm8. That value was
seen as result when interpolating between 2 and 29 as in the following
expressions:

result = 2 / 31.f  * 0.5f + 9 / 31.f * 0.5f
unorm8 = 255.f * result

While the exact value if calculated with rational numbers is 127.5f,
the floating poing math leads to a value that is sligthly below 127.5f.

The RTL handled the value as a tie because of being closer than 1/256
to 127.5 and rounded up to 128. TBoxEmu on the other hand did not have
any epsilon added and therefore did round it down to 127. I have now
introduced an epsilon on TBoxEmu as well, so both implementations
apply now the same rounding rule.

Finally, when everything was fixed and matched between RTL and TBoxEmu,
and I could verify a the mismatching cases by hand against DirectXTex, I
enabled the VK_FORMAT_BC1_RGB_UNORM_BLOCK and
VK_FORMAT_BC1_RGBA_UNORM_BLOCK formats on decompress_tb.cc.

This patch was needed for RTLMIN-1973.