[esperanto-tools-libs] [RT] Remove the et_bootrom.h interface.
The functionality for specifying the path and loading the bootrom
has been rolled inside the Device class.

Use that instead of the old C API where we were manually updating the
pointers.

Cleanup RuntimeStub to pass the path to the bootrom to the Device class
and remove the logic for reading the bootrom.

Remove the old et_bootrom.* files

fixup! [RT][SW-261] Enable using the device manager inside SysEmuFunction.