[esperanto-tools-libs] [RT] Add support in Device to hold multiple target devices.
This commit:
* Adds support for allocating and holding target devices of different
subtypes (e..g sysemu-et-rpc etc)
* Add support for specifying and loading the bootrom file.
* Adds explicit init and resetDevice functions so that we can decouple
device init/reset of the class constructors. Both init/reset can
generate errors and for now we want to be able to explicitely handle
them and avoid to have to throw exceptions inside the constructor.
Also no exception can be thrown inside the destructor.

--- Logging.h add RTERROR helper macro

--DeviceTarget:
* Add helper function for querying the type of device to
create
* Add function to check if the underlying device is "alive" e.g. check
if the simulator is running.

-- MemoryManager:
* Add explicit init and deInit function and decouple that functionality
from the constructor and destructor