[esperanto-tools-libs] [RT] TargetCardProxy simplify simulator launch.
This commit instead of using a thread and mutex/locks trying to
keep track of the thread that launches sysemu, instead it launches
the simulator and waits on a future that blocks until the function
that launches the simulator has returned. Internally we wait for the
simulator to exit. SysEmu will exit either by encountering an error
or by sending it an explicit shutdown command that forces it to
exit gracefully.