[esperanto-tools-libs] [RT] Disable building old kernels and bootrom.
The old kernels and bootrom are not being used disable building
them and remove from the runtime support for launching them.

Current ifdef out any code we wanto to disable, in the future we should
remove that code and establish an API for a library to register itself
with the runtime so that it can be loaded on the device.