[esperanto-tools-libs] [BUILD] Add support for building etsoc-glow using CMake
This commit adds a target etsoc-glow that build the glow_fork
version of glow using CMake. This target has similar functionality
as the build.sh script

Also we stop sourcing the runtime folder and we start building it
as an external project. As part of that we are moving under the runtime
folder the gtest dependency and fix the runtime build to configure
using the runtime folder directly.

Finally when we build the runtime to get linked with Neuralizer
we force it to generate Makefiles and we force the number of threads
to use to compile.

Also we stopped building sw-sysemu for host-sw in build.sh