[esperanto-tools-libs] [RT] Add ELF support helper classes.
This commit tries to clean-up the logic inside registry.h and
registry.cc. It creates a base class that will hold the information
for an ELF. In the future this class will be subclassed for handling
both the case of a kernel-elf and a firmware-elf file.

Tests are going to be added incrementally. For convenience we are also
merging one of our kernel elf files to enable respective testing.