[esperanto-tools-libs] [RT] Change moduleLoad to take elf name and path.
This commit:
* Modified the Module class to be composed of the KernelELFInfo
and hold the raw data of the elf. This way upper layers pass to the
runtime the module's name and path to the elf-file ad they do not need
to hold the raw data of the elf file.
* The interface to Device::moduleLoad has been modified to be the
name and path of the ELF. We are using internally the changes to the
Module class to hold the ELF's information and we remove the call
to the "parse_elf" function that has been replaced by the functionality
introduced in the KernelElFINfo class