[esperanto-tools-libs] [RT] Remove from EtAction shared_pointer semantics.
EtAction was implementing its own version of "smart" pointers
keeping a reference count that was excplitely increased and decreased.

This is error prone.

Remove the reference count and the related API replace all instances of
EtAciont pointers with smart pointers. Convert the list of actions in
the Stream to be a list of shared pointers.