[esperanto-tools-libs] [RT] Create basic Command and Respose types of the DeviceAPI
Create basic Command and Response types of the DeviceAPI.
Future types will be derived from these 2 basic types, which are going
to be used for replacing the existing Action* classes in the Commands.h
file.
Each Command type will be paired with its respense type. A Command
internally holds a std::romise<Response> object that will be used to
generate an std::future<Response> object that the user can create
so that they can wait/block until the value of the Repsonse is set.

Moving forward these 2 primitives will allow us to create multiple
threads of executing in the runtime (depending our needs). In the bare
minimum for each Device we are going to have a separate execution thread that is
going to read/write/poll/lselect the device file-descriptor of the
Esperanto character-device.

Per thread of execution we are going to have a Command Queue we are
going to submit Command objects. Once the operation of each Command
is completed we are going to update the results of the respective
Response object.