[sw-sysemu] [ZSIM] Initial commit
ZSIM is an extremely light-weight and simple checker for fast ZEBU runs.

ZSIM is not a replacement of CoSIM, and it is not meant to check state to the
depth that CoSIM does. Its main purpose is to be very simple and fast, at the
cost of permiting some escapes.

Doing 'et-dvrun $MINION_REG --uarg zsim' enables ZSIM instead of CoSIM.

ZSIM is still work-in-progress. This initial version only works with tests
that do not use synchronization primitives or share state among threads, and
simply validates the PC of each Minion retired instruction, trapped exception,
or trapped interrupt against BEMU.