[test-compute-kernels] [SW-358][BUILD] Convert device-fw to a package
This commit adds support for exporting the different targets
that device-fw  build as a cmake package that other downstream
dependencies can find using the "find_package" mechanism.

The commit also changes the install targets separating the elf
files in the bin folder and the hex files in the lib folder.
This is "tentative" as this will not be the "right" place where
the elf files would be installed. Also the hex files will not be
released.

Also adding the helper kernels to the list of targets we release