[esperanto-tools-libs] [RT][BUILD] Remove abseil as a submodule.
This commit removes abseil as a submodule. Instread it uses cmake
to explicitly clone and build the library and runtime build time.
This change has been done to mitigate some CI issues with the submodule
but it is also the direction we are going to follow in the runtime
for any public external dependencies:
Use cmake find_package etc to find the library and it if does not
exit then build it from source as necessary.

This is necessary to handle the 2 build modes of our code we need to
support in the future:
1) Build runtime and all dependencies and relecase them as artifacts
2) Build runtime consuming any library dependencies