[esperanto-tools-libs] [SW-358] Make device-fw a build dependency of the runtime
Add call find_package(esperanto-fw) in the runtime. If an
installation of espranto-fw is found in the system (or passed in the
build options) then:

1) Related tests that require device-fw are enabled and the
tests find the paths of the different elf image from the cmake
package information.
-- Disable downloading old device-fw images from artifactory that
were used for those tests.
2) Additional command line options have been added in DeviceFW.cc
where we can specify the path of each individual *Minion ELF
-- The test_load_device_fw.cc has been modified to read the values
of those command-line options
3) The runtme/tests has been extended to support passing additional
command line options to the tests, to support the above use case.
4) ParseCommandLineOptions helper function has been introduced for
   parsing the abseil command flags.