[esperanto-tools-libs] [RT] Remove DefineDevMem calls form the code base.
RPC calls for "defineDevMem" are an artifact of SysEmu where
we needed in the past to "tell" the simulator the size and properties
of the memory region to create. This is not needed any more as it will
allocate memory no demand.

Remove from the Device class hierarchy any such references as this is
functionality that is not needed for the real device.