[esperanto-tools-libs] [SW-363][RT] Implement support for the Mailbox devices
This commit adds preliminary integration with the PCIE mailbox
devices. Also it implements the launch function of the PCIeDevice
class using the MailBox device.

Currently we craft a specific message that device-fw expects. The
definitions of this message come from device-fw.
A macro and boolean are added in ProjectAutogen.h.in that are true
iff the runtime has found the espernato-fw package and can consume
the related headers released by device-fw.

When loading an ELF if no entrypoint functions are found then assume that the kernel
offset is the load address off the zero segment. This handles the
case of the device-fw kernels that are compiled such that the launch-pc
is the load address of segment-0. Given the current implementation
we copy the whole elf-file on the device and the launch PC is the
copy-address-of-elf-file + segment-0-file-offset. This approach is a
HACK that will need to be fixed soon.