[esperanto-tools-libs] [RT][SW-547] Handle loading kernels that have different "load" adddress
This is an intermediate step for correctly loading the kernels correctly
on the device.
Currently we need to handle 2 cases:
1) One where kernels are compiled with low load addresses outside
the DRAM space
2) Another where the kernels are compiled for fixed load addresses
within the DRAM space.

For the first case we need to "relocate" the kernel at the right
place, for which case we "malloc" memory that is forced at the right
place. The second we need to load the segment at the specified
load address.

This commit handles the two cases and adds a test case for such a
sample-kernel case.