[esperanto-tools-libs] [SW-586][SW] Handle correctly loading ELFs with multiple segments
This commit removes the restriction where we could not load on the
device ELF files that conntain multiple segments. This commit
removes this restriction and loads any number of ELF segments on the
device.

Still though code-relocation is not handled correctly. And this
change is done to unblock further debugging.

In this commit we are also adding additional tests that exercise this
case.