[et-driver] Make IOCTLs check minor device (SW-642, SW-641)
This commit reduces the number of IOCTls and makes them check for
appropriate minor device numbers/types. For example, the individual reset-
mailbox IOCTLs are megerd into a single IOCTL that resets the mailbox
that corresponds to the minor device on which it was called and returns
an error for an inappropriate minor device. Similar improvements and
checks are implemented for other IOCTLs. The changes are tracked in Jira
SW-642. Also resolves SW-641.

Address review comments

This commit removes some TODO comments (and fixes the reason for them)
and declares a variable in a local scope, where it belongs.