[esperanto-tools-libs] [SW-1023] Create SysEmu run folder
Add an option in the runtime to specify the sysemu runtime folder
if not specified then we are going to create a random folder in the
CWD that is suffixed with the PID of the executing process and a
random number given that the runtime+sysemu can be launched multiple
times in the lifetime of a process.

In that folder we save the sysemu uart logs, and the sysemu stdout+stderr
log. We also create a file where were redirect the sysemu stdout and stderr.

We also simplified how we select the GRPC socket file name to use.
Avoid creating a random temporary file and the associated complexity.
Create a namespace abstract unix socket using the "unix:@NAME" syntax.
Even though this form does not allow us to avoid creating a file,
it does create a file in the local running directory.