[esperanto-tools-libs] [SW-1083][RT] Initialize the EmuMailbox by default
This commit is initializing the emulated mailbox by default
as part of the device-init. Initialization requires taking the mailbox
to the ready state and also "resetting it".

To enable the above we have created a postFWLoadInit step that executes
once we have loaded and booted device-fw on the simulator. This a
temporary step that does not correspond necessarily with what the real
device would do.

In the process we have also adapted the tests to correctly run the
steps that have now moved inside postFWLoadInit or call the init()
function where it is currently possible

As a side-effect we also specify the --fake_fw path and set it for
the tests where necessary