[esperanto-tools-libs] [SW-1085] Do not build runtime dependencies if exist in Docker
This commit using the cmake find_package functionality to skip build
runtime library dependnecies if they already exist in the docker image
This is in an effort to reduce the build time.

We still need though to build all dependencies when we release artifacts
to etsoc so that they have all library dependency source, and they do
not have to build with docker.

For this reason the BUILD_DEPENDENCIES option has been added, that when
true we are going to build all library dependencies as normal.

Also a number of fixes has been done as part of this process to make
runtime a "proper" cmake package that correctly captures all
dependencies