[esperanto-tools-libs] [SW-210][RT] Use Doxyfile to generate the doxygen documentation
Use a doxyfile for generating the documentation, instead of of
using "breaths" approach. This way we also generate the doxygen
original documentation for comparision.

We are also introducing a couple of empty for now documentation files
for other topics that will be populated in the future. Test for
now that we can easily include other files.