[device-api] [SW-1191] Introduce the YAML specification of DeviceAPI
This commit introduces the infrastructure to describe in YAML
the DeviceAPI

The intent is to describe in YAML and valdiate its correctness using
json-schema the different types of messages that the DeviceAPI
will contain.

From the YAML-schema we will auto-generate code to different
target languages.
We also want the schema to be consumable by the host Runtime tracing
infrastrucuture so that we can automate converting the Device-API
trace-messages from binary to the target format in the host Runtime
supports (e.g. protobuf)

This commits contains
* The jsonschema specification of the device-api
* The YAML specification of a number of DeviceAPI messages
* Simple code-autogeneration, where we generate a C header that
can be consumed by device-fw