[esperanto-tools-libs] [RT][SW-1191] Consume the DeviceAPI schema
This commit consumes the DeviceAPI schema, both the jsonschema specification
and the YAML DeviceAPI schema.
The above files are released as build artifacts of the device-api
project. The commit modifies the etrt-trace.yaml and tracing-spec.schema
to be configured and cmake-configuration time and point to the
absolute paths of the DeviceAPI files they need to include.

This is a necessary step so that we can generate-code in the
runtime and convert device_api trace events to protobuf in the future