[esperanto-tools-libs] [SW-1281] Add enum and struct support in the tracing schema
In an effort to covert any of the device-api messages or trace
events in the format that the protobuf-trace and the runtime
tracing infrastructure will support, introduce enum and struct
support.

This will enable us to convert the enum and struct types specified
in the device-api in tracing messages.

The goal is to record any device-api message in the runtime trace