[device-api] [SW-1281] Introduced pairing information across messages.
Introduce an additional field in the Message definition called
"PairedMessage" that allows us to describe the relationship between
Command and Response messages. For events the PairedMessage has to be
done.

Add check in device_api_codegen to make sure that the name of the
paired message is a valid one