[esperanto-tools-libs] [SW-1368] Add ELF path to the CodeModule constructor.
Modify the Module constructor to take as arguments the module's
name and path as strings.

Move the module_count to be a static member of the Module class.

Modify the readELF function not to take the path as an argument but
instead use the class member information

Fix the ELFINfo constructor as well and all consumers of the Module
class.
Also Fix related tests