[device-minion-runtime] [SW-1308] Record the mtime timestamp in the DeviceAPI commands/replies
This commit records the mtime in the DeviceAPI commands and replies,
that is to be recorded by the runtime in the runtime-trace and be
used for performance/latency analysis.

In the process the constness has been removed by the buffers being
passed in the handle_message_* functions, so that those functions
can modify the body of the received command with the appropriate
information.