[esperanto-tools-libs] [SW-1175] Add test that records a string and bytes in the trace
Add a Test fake module in the tracing infrastructure where we
introduce two "fake" functions that record a string and bytes.

Introduce a small unit-test that records a string and a vector.
The test exercises the interface and code-generation, but not
if the information is recorded correctly. this is a little much
harder.