[esperanto-tools-libs] [SW-1279]  Create a thread that reads the mailbox for Response messages.
This commit:
* Introduces a Command callback registration mechanism. This registry
will execute the callback for a specific Command once the respective
Response has been received from the mailbox

We have 2 separate categories of messages the FW-messages and the
DeviceAPI messages. The DeviceAPI do record exactly which is the
ID of the Command Runtime-object while the FW-messages do not.
As such it is straightforward when looking in the DeviceAPI
reponse to idenitfy the Command that requested it.

For this reason we store the FW-messages in a std::queue and we expect
replies from FW to come in order, and DeviceAPI commands are saved
in a std::map that we look up based on the CommandID.

* Introduces a Device Thread that consumes all Commands submitted
by all threads and serialized in a common queue the executes them
in order.

* Introduces a thread that reads and blocks on the mailbox until a
new Response or Event arrive. If a new Response has arrive then
we execute the Command Callbacked registered for the respective
Command. The callback sets the Response value in the Command's promise
object, thus allowing the std::future::get() to return and unblock
the thread that is waiting for the response of a specific Command.

* The existing Command execute functions have been modified to register
a callback with the Device.

* Misc other fixes