[esperanto-tools-libs] [SW-1279] Update simulator-API to enable async host-interrupt
Update the simulator-api and pull changes in the sim-api-server
that handle asynchronously requests for host-interrupt.

This change will spawn a new thread that will wait for the device to
raise an interrupt from the device to the host. The mail thread
will be unblocked to handle follow-up requests and not be blocked
on the host-interrupt request.

This will allow us to modify the runtime to have 2 threads:
1) One that sends commands to the device
2) One that blocks on the mailbox-read for Responses and Events
to be generated back from the device. The mailbox-read blocks
waiting for an interrupt to be raised from the device to the host.