[esperanto-tools-libs] [RT] Fix incorrect gtest fixtures and device init/deinit
The DeviceFWTest was miss-used in the related tests and resulted
in the SetUp and Teardown function not to be symmetrical: do the
initialization and reset of the Device.

Inttroduce MailboxEmuTest  fixture that is to be used by the
mailbox emulation tests, that operate only at the GRPC target
device.

Fix tests as necessary.