[esperanto-tools-libs] [SW-1566] Fix UberKernel runtime integration and add test
As a temporary hack use the MasterShire bit to pass the informaiton that
we are launching an uber-kernel. Fix the interfaces as necesasry

Create an integration test that consumes the uber-kernel
defined in the device-firmware. This is a 2 layer kenrel that writes
twice the "beef" pattern.

The test calls the uberkernel and checks if the "beef" has been written
in the allocated tensors.