[esperanto-tools-libs] [RT] Clean-up uses of the DeviceAPI
1) [RT] Remove the fw_command support and ReflectTest support

Use the newly introduced reflect_test_cmd DeviceAPI command to
implement the reflect mailbox test.

Remove the existing ReflectTest command/response and the serial
queue that used to support the firmware-messages that we are
dropping support for. This will eliminate the need to include
any device-firmware hearders.

2) [SW-1321][SW-1382] Remove the explicit LaunchCommand class use DeviceAPI

Remove the LaunchCommand class and move the functionality
inside the KernelLaunch classes of the Kernel and the UberKernel

Implement blocking and non-block kenrel-launch functions. Refactor
the code and create helper functions where necessary