[esperanto-tools-libs] [SW-1458] Static assignment of L2 scp entries
Reworked the scratchpad layout for A and B operands in preparation
for the weight decoupling.

Now there is a 0.5 Mb static allocation for A, where either A fits
or it gets spilled from the scratchpad.

For B there is a fix allocation of 1 Mb and similarly to what it
happens for A, either B fits in that allocation or it gets spilled out
of the scratchpad.

For the generic checkL2SCPSize implementation in LoadParametrize we
also force that when one operand (either A or B) gets spilled the
other either is not allocated in the SCP or it gets spilled. This
constrains the scratchpad allocation in such a way that we never see
the case that A is allocated in the SCP but B is not, or the other way
around (B is allocated in the SCP but A is not).

Also for the case of using a parallel factor other than 0, all the A
operands are allocated together. The same is done for the B operands
that are also packed together. This is in constrast with the previous
layout where A and B from a parallel operation where allocated in
consecutive memory addresses.