[device-bootloaders] [SW-1839] Remove the ServiceProcessorROM submodule
This removes the ServiceProcessorROM submodule and building
it as part of device-fw. This is a necessary step for
sanitizing device-fw and releasing it to the client.

Scripts:
otp_skip_vault.sh
src/build_sp_sram_image.s

Are still referencing the ServiceProcessorROM submdoule but are
left for reference here as we have not migrated their flow
inside the sw-platform