[device-minion-runtime] Remove the uplfash.sh script and build folder.
The build folder should not be part of the repo.
The destination of the build folder can be anywhere in the filesystem
and outside the repo.

Also there is no need to have a script to copy artifacts from the
user to the synopsys cloud. This does not follow the overall build
flow and zebu regressions. It was a hack that Doug had in place.