[device-minion-runtime] [SW-1961] Make the kernels folder a self-contained cmake package
In preparation for extracting the kenrels out of device-fw
convert the kernels folder to a self-contained cmake package
that can be built independently from within sw-platform.

We have replicated the Toolchain-gcc-riscv.cmake and layout.ld
files. The rest of build dependnecie that we used to shared in device-fw
as consumed as build artifacts exported by device-fw inb the
esperanto-fw package

An option has been added in device-firmware to disable building
the kernels as part of the device-fw build target. This will
be used in sw-platform when we switch in building the test-kernels
from with device-fw to the new self-contained package.