[esperanto-tools-libs] [SW-2006] Remove the elfio submodule
This commit removes the elfio submodule and converts it to a build
dependency: as part of building elfio cmake ExternalProject will
clone the github repo and chekcout a specific commit.

After this step the runtime has no internal Esperanto submodules
in its source and can be extracted to a separate repo.