[device-minion-runtime] [SW-2010] Use the sw-platform component builder job to build device-fw
This commit moves the checkin job to use the sw-platform component
builder to build device fw

Given that Device-FW is not self-contained and generates a number of
outgoing dependencies that other components consume, in the
checkin build we are not only building device-fw but the whole
flash-image for now to make sure we are not breaking the bootrom
build. This should be revisited in the future.

This change is necessary so that we can remove the infra-tools submodule
from this repo