[device-minion-runtime] [SW-2022] Remove unused git submodules, prepare for client release
This commit removes the followsing submodules:
* infra_tools
* firmware-tools
* device-api
This allows us to sync the repo directly in github

After this  the repo is not self-contained any more.
It will need to be buillt as part of sw-platform where
it will consumed any input build depednencies (device-api, fw-tools)
and generate downstream build artiacts.

Unused build scripts, and CI infrastructure has been remove.
This also removed the et-sw-device-fw docker image.