[device-minion-runtime] [SW-2342] Remove code that belonds to the SP-ROM
This commit stops installing as part of the fw-helpers a number of
headers, the vaultip headers/library that have been moved inside the
SP-ROM repo.
The SP-ROM will be releasing the headers as libraries that the
device-bootloaders are giong to consume

The files have not been deleted and are kept around so that the
it is easy to do see the history.