[esperanto-tools-libs] [SW-3188] Remove include path pollution
This commit:
* Creates a object library for the C-API

* Removes from the top-level src/CMakeLists.txt file the extensive
list of include paths that had been added and was resulting in
the install path to be included as well. As a result the code
was compling against the "old" headers of the runtime copied
in the install folder, instead of consuming the correct version
in the source tree.