[esperanto-tools-libs] [SW-3254] Run SysEMU in an infinite loop
Set the sysemu max_cycles to UINT64_MAX int so that sysemu
runs in an infinite loop. This is needed to enable correctly
the SimulatorAPI asynchronous mode where SysEMU needs can handle
in parallel executing both instructions on the cores and handling
of SimulatorAPI requests.

Ignore the command-line value, do not remove the command-line
option for now, as that will cause test-framework changes.