[esperanto-tools-libs] [SW-3340] Introduce the Basic and Linear Memory Allocator classes
This commit introduces a first implementation of the Basic and
Linear Memory allocator classes.

The Linear Memory allocator is a simple implementation of a memory
allocator that mantains 2 lists: a free and an allocated one.

So far we have implemented a simple interface that does malloc/free
and can print basic stats from the allocator