[esperanto-tools-libs] [SW-3408] Implement the bidiretional memory allocator
This commit implements the bidirectional memory allocator that is
being used to hold the constant and placeholder tensors of the system.

The BasemMemoryAllocator interface has been modified as the
bidirectional allocator has 2 separate malloc function to specify
the direction of allocation.

A set of unit-tests that exercise different patterns of allocations
and frees has been added.