[device-api] [SW-3199] Introduce the privileged Device API
This commit introduces the set of Device API RPC calls that
belong to the privilged level and should be executed only
by Linux kernel-space processes, and not by a user-space application.

This commit introduces the command/response pair to fire a DMA
on the Device, after the DMA descriptors have been sent over,
and wait for the response that reports the status of the DMA channel.