[device-api] [SW-3199] Use the enum type name in the schema, not generated c-type
This commit fixes a bug where when a message or a struct had a member
field of type-enum, they would reference the type by the generated
codes c-name (e.g. dev_api_kernel_launch_error_e) and not by the
original type name specified in the API (e.g.
dev_api_kernel_launch_error_e)

This commit addresses this problem. It introduces the Enum helper
class in the code-generator that returns the c_type_name of a given
type.

Fix the rpc-calls.yaml to reference the correct type names.

We also fix a number of incorrect string comparisons in the
generator code.