[device-api] [SW-3199] Add ability to specify the range of the message ids
This commit adds the ability to specify in the API schema the range
of values we want the message-IDs of our API to have. This information
will be used when auto-generating C/C++ enumerations with the message
ids.
This will enable us to define the privileged and non-privileged
API messages and ensure that they have separate enumerations.