[esperanto-tools-libs] [SW-3199] Accomodate the introduction of the non-privileged DeviceAPI
The DeviceAPI has been split into 2-segments the non-privileged and
the privileged one (it is effectively two APIs)

The host runtime is allowed to execute only the non-privileged set
of DeviceAPI commands.

Include the device_api_cxx_non_privileged.h header, where we have
separated the CXX part, and fix all the namespace and type renaming.