[esperanto-tools-libs] [SW-3532] Use the new memory Allocators in the MemoryManager.
Use the new memory allocators (Linear and Bidirectional) in the
MemoryManager by creating an instance of MemoryManagerInternals
object.

As a first step we want to maintain the same interface to
Glow-Backend. Use the Linear allocator to allocate any
code and data is always going to be allocated from the
from of the bidirectional allocator.

Create a Code Region that is currently 4GB and the rest of
DRAM is the data region.

Still return absolute addresses.

Modify the malloc/emplace interfaces to return a tuple of
BufferId and Offset for each allocated buffer. The offset
is effectively currently the absolute address inside the
SOC DRAM