[esperanto-tools-libs] [SW-3532] Ensure that buffer meta-data is cache aligned
Change the buffer meta-data size to be a multiple of a cache line.
This will ensure that the buffer allocated will be cache-line
aligned.

Also modify the BufferOFfsetTy and BufferSizeTy types to be signed
integers, to avoid integer underflow bugs and ease comparisons

Fix the tests where the above changed increased the size of the
metadata